requirements.txt
python-telegram-bot==20.7
openai==1.12.0
sqlalchemy==2.0.25
pydantic==2.6.1
python-dotenv==1.0.1
pytz==2024.1
pytest==7.4.4
pytest-asyncio==0.23.3
aiohttp>=3.9.0
google-api-python-client==2.114.0
google-auth==2.27.0
google-auth-httplib2==0.2.0

